+++ index.html
@@ -231,6 +231,6 @@
 </template>
 
 <!-- JavaScript-->
-<script src="./js/main.js"></script>
+<script src="./js/main.js" type="module" defer></script>
 </body>
 </html>
